+++ R/header-transform.R
@@ -26,8 +26,13 @@ merge_header <- function(x, merge = TRUE) {
 }
 
 transform_header <- function(
-                             x, sep = "[_\\.]", theme_fun = flextable::theme_booktabs,
-                             .fill = FALSE, .merge = FALSE, ...) {
+  x,
+  sep = "[_\\.]",
+  theme_fun = NULL,
+  .fill = FALSE,
+  .merge = FALSE,
+  ...
+) {
   header <- names(x$header$dataset)
 
   mapping <- data.frame(original = header, stringsAsFactors = FALSE) %>%
@@ -37,6 +42,15 @@ transform_header <- function(
       sep = sep, fill = "right", remove = FALSE
     ) %>%
     fill_header(.fill)
+
+  if (is.null(theme_fun)) {
+    default_theme <- flextable::get_flextable_defaults()$theme_fun
+    theme_fun <- if (is.function(default_theme)) {
+      default_theme
+    } else {
+      getNamespace("flextable")[[default_theme]]
+    }
+  }
 
   x %>%
     flextable::set_header_df(mapping, key = "original") %>%
@@ -50,6 +64,10 @@ transform_header <- function(
 #' @param x A `flextable` object`
 #' @inheritParams tidyr::separate
 #' @inheritParams flextable::flextable
+#' @param
+#'   theme_fun A flextable theme function.
+#'   When `NULL` (default), the value is resolved by
+#'   `flextable::get_flextable_defaults()`.
 #' @param ... Passed to `theme_fun`
 #'
 #' @examples
@@ -58,7 +76,11 @@ transform_header <- function(
 #'   separate_header()
 #' @export
 separate_header <- function(
-                            x, sep = "[_\\.]", theme_fun = flextable::theme_booktabs, ...) {
+  x,
+  sep = "[_\\.]",
+  theme_fun = NULL,
+  ...
+) {
   transform_header(
     x,
     sep = sep, theme_fun = theme_fun, .fill = FALSE, .merge = FALSE, ...
@@ -76,7 +98,11 @@ separate_header <- function(
 #'   span_header()
 #' @export
 span_header <- function(
-                        x, sep = "[_\\.]", theme_fun = flextable::theme_booktabs, ...) {
+  x,
+  sep = "[_\\.]",
+  theme_fun = NULL,
+  ...
+) {
   transform_header(
     x,
     sep = sep, theme_fun = theme_fun,